requirements.txt
azure==4.0.0
plotly
sklearn
requests
xmltodict
numpy
Flask==0.12
Pillow==3.1.2
tensorflow
keras
arrow==0.12.0
grpcio==1.27.2